fix map images
