+++ server/src/parse-map-packs.ts
@@ -14,36 +14,6 @@ import {
   assertParentMap
 } from "@tsconline/shared";
 
-/**
- * Finds all map images and puts them in the public directory
- * For access from fastify server servicing
- */
-export async function grabMapImages(datapacks: string[], destination: string) {
-  const imagePaths = await grabFilepaths(datapacks, assetconfigs.decryptionDirectory, "MapImages");
-  const compiledImages: string[] = [];
-  try {
-    // recursive: true ensures if it already exists, we continue with no error
-    await fs.mkdir(destination, { recursive: true });
-    await pmap(
-      imagePaths,
-      async (image_path) => {
-        const fileName = path.basename(image_path);
-        const destPath = path.join(destination, fileName);
-        try {
-          await fs.copyFile(image_path, destPath);
-          compiledImages.push(`/${destPath}`);
-        } catch (e) {
-          console.log("Error copying image file, file could already exist. Resulting Error: ", e);
-        }
-      },
-      { concurrency: 5 }
-    ); // Adjust concurrency as needed
-  } catch (e) {
-    console.log("Error processing image paths for datapacks: ", datapacks, " \n", "With error: ", e);
-  }
-  return compiledImages;
-}
-
 export async function parseMapPacks(datapacks: string[]): Promise<MapPack> {
   const mapInfoPaths = await grabFilepaths(datapacks, assetconfigs.decryptionDirectory, "map-packs");
   const mapInfo: MapInfo = {};
